[gaim-migrate @ 9677] " Fixes a few warnings on 64bit machines. Also fixes yahoo auth on 64bit machines which would have taken me awhile if marv didn't point me in the right direction. I've applied to my local 32bit copy and everything seems to be working fine." --Gary Kramlich
committer: Luke Schierer <lschiere@pidgin.im>

--HG--
extra : convert_revision : a18b28ba392e3a545cc64ac2b52c8562ddbf1a3b

+++ src/protocols/oscar/oscar.c
@@ -5626,7 +5626,7 @@ static int oscar_send_chat(GaimConnection *gc, int id, const char *message) {
 	char *buf, *buf2;
 	char *charset = NULL;
 	int encoding;
-	int len;
+	gsize len;
 
 	if (!(conv = gaim_find_chat(gc, id)))
 		return -EINVAL;
